Move audio types into pipeline
