+++ compositor_render/src/types.rs
@@ -1,56 +1,6 @@
 use std::{collections::HashMap, fmt::Display, sync::Arc, time::Duration};
 
 #[derive(Debug, Clone)]
-pub struct AudioSamplesSet {
-    pub samples: HashMap<InputId, Vec<AudioSamplesBatch>>,
-    pub start_pts: Duration,
-    pub length: Duration,
-}
-
-impl AudioSamplesSet {
-    pub fn end_pts(&self) -> Duration {
-        self.start_pts + self.length
-    }
-}
-
-#[derive(Debug)]
-pub struct OutputSamples(pub HashMap<OutputId, AudioSamplesBatch>);
-
-#[derive(Debug, Clone)]
-pub struct AudioSamplesBatch {
-    pub samples: Arc<AudioSamples>,
-    pub start_pts: Duration,
-    pub sample_rate: u32,
-}
-
-impl AudioSamplesBatch {
-    pub fn end_pts(&self) -> Duration {
-        self.start_pts
-            + Duration::from_secs_f64(self.samples.len() as f64 / self.sample_rate as f64)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum AudioSamples {
-    Mono(Vec<i16>),
-    Stereo(Vec<(i16, i16)>),
-}
-
-impl AudioSamples {
-    pub fn len(&self) -> usize {
-        match self {
-            AudioSamples::Mono(samples) => samples.len(),
-            AudioSamples::Stereo(samples) => samples.len(),
-        }
-    }
-
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-}
-
-#[derive(Debug, Clone)]
 pub struct Frame {
     pub data: YuvData,
     pub resolution: Resolution,
@@ -151,10 +101,4 @@ impl Resolution {
     pub fn ratio(&self) -> f32 {
         self.width as f32 / self.height as f32
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AudioChannels {
-    Mono,
-    Stereo,
 }
